Redux

+++ backend/groupsCRUD.py
@@ -1,11 +1,11 @@
 from fastapi import HTTPException, status
 from sqlalchemy import update
 from sqlalchemy.orm import Session
-import models
-from schemas import groups
+import models, schemas
+import tasksCRUD
 
 # Create
-def create(db: Session, group: groups.GroupCreate):
+def create(db: Session, group: schemas.GroupCreate):
     exists = db.query(models.Group).filter(models.Group.name == group.name).first()
     if exists:
         raise HTTPException(
@@ -13,7 +13,7 @@ def create(db: Session, group: groups.GroupCreate):
             detail="Existe un grupo con este nombre"
         )
     else:
-        db_group = models.Group(name=group.name, description=group.description)
+        db_group = models.Group(name=group.name)
         db.add(db_group)
         db.commit()
         db.refresh(db_group)
@@ -28,7 +28,7 @@ def get_all(db: Session):
     return db.query(models.Group).all()
 
 # Update
-def refresh(db: Session, group_id: int, group: groups.GroupCreate):
+def refresh(db: Session, group_id: int, group: schemas.GroupCreate):
     db_group = db.query(models.Group).filter(models.Group.id == group_id).first()
     if not db_group:
         raise HTTPException(
@@ -36,7 +36,6 @@ def refresh(db: Session, group_id: int, group: groups.GroupCreate):
             detail="Group not found"
         )
 
-    db_group.description = group.description
     db_group.name = group.name    
     db.commit()
     
@@ -46,9 +45,15 @@ def refresh(db: Session, group_id: int, group: groups.GroupCreate):
 def delete(db: Session, group_id):
     db_group = get_one(db=db, group_id=group_id)
     if db_group:
+        db_task = tasksCRUD.get_all(db=db, group_id=group_id)
+        if(db_task):
+            raise HTTPException(
+                status_code=444,
+                detail="El proyecto tiene actividades, eliminelas para poder eliminar el proyecto"
+            )
         db.delete(db_group)
         db.commit()
         db.close()
-        return "Group deleted"
+        return "deleted"
     else:
         raise HTTPException(status_code=404, detail=f"group with id {group_id} not found")
